GhidraLLM/src/main/python/requirements.txt
langchain==0.1.0
google-generativeai==0.3.2
protobuf>=3.19.5,<5.0.0dev
psutil==5.9.8
setuptools==68.0.0
wheel==0.37.1 